crop1/backend/requirements.txt
Flask==2.2.5
Flask-Session==0.4.0
werkzeug==2.2.3
itsdangerous==2.1.2
fpdf==1.7.2
python-docx==0.8.11
scikit-learn==1.2.2
pandas==2.1.2
numpy==1.26.2
gunicorn==20.1.0
